PS-87 Beginning of locks

+++ src/server/Persistify.Server.Persistence.LowLevel/Streams/SelfManagingFileStream.cs
@@ -12,6 +12,7 @@ public class SelfManagingFileStream : Stream
     private CancellationTokenSource? _cancellationTokenSource;
     private readonly object _lock;
     private readonly TimeSpan _idleFileTimeout;
+    private bool _fileExists;
 
     public SelfManagingFileStream(
         TimeSpan idleFileTimeout,
@@ -21,13 +22,18 @@ public class SelfManagingFileStream : Stream
         _filePath = filePath;
         _lock = new object();
         _idleFileTimeout = idleFileTimeout;
+        _fileExists = File.Exists(_filePath);
     }
 
     private void EnsureFileStreamOpen()
     {
         lock (_lock)
         {
-            _fileStream ??= new FileStream(_filePath, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.None);
+            if (!_fileExists)
+            {
+                throw new FileNotFoundException($"File '{_filePath}' does not exist.");
+            }
+            _fileStream ??= new FileStream(_filePath, FileMode.Open, FileAccess.ReadWrite, FileShare.None);
             CancelCloseFileTask();
             ScheduleCloseFile();
         }
@@ -147,11 +153,24 @@ public class SelfManagingFileStream : Stream
         base.Dispose(disposing);
     }
 
+    public void CreateUnderlyingFile()
+    {
+        if(_fileExists)
+        {
+            throw new InvalidOperationException("File already exists");
+        }
+        File.Create(_filePath).Dispose();
+    }
+
     public void Delete()
     {
         CancelCloseFileTask();
         _fileStream?.Dispose();
         _cancellationTokenSource?.Dispose();
+        if (!_fileExists)
+        {
+            throw new InvalidOperationException("File does not exist");
+        }
         File.Delete(_filePath);
     }
 }
